@@ -2,7 +2,7 @@
     "type": "project",
     "license": "proprietary",
     "require": {
-        "php": "^7.2.5",
+        "php": "^7.4",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "symfony/console": "5.0.*",
@@ -19,9 +19,9 @@
         "sensio/framework-extra-bundle": "^5.1"
     },
     "require-dev": {
+        "doctrine/doctrine-fixtures-bundle": "^3.3",
         "symfony/maker-bundle": "^1.14",
-        "symfony/var-dumper": "^5.0",
-        "doctrine/doctrine-fixtures-bundle": "^3"
+        "symfony/var-dumper": "^5.0"
     },
     "config": {
         "preferred-install": {
