@@ -6,6 +6,7 @@ use App\Repository\ApiTokenRepository;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
+use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 use Symfony\Component\Security\Core\Exception\AuthenticationException;
 use Symfony\Component\Security\Core\Exception\CustomUserMessageAuthenticationException;
 use Symfony\Component\Security\Core\User\UserInterface;
@@ -14,11 +15,17 @@ use Symfony\Component\Security\Guard\AbstractGuardAuthenticator;
 
 class ApiTokenAuthenticator extends AbstractGuardAuthenticator
 {
-    private $apiTokenRepo;
+    private ApiTokenRepository $apiTokenRepo;
+    /**
+     * @var UserPasswordEncoderInterface
+     */
+    private UserPasswordEncoderInterface $passwordEncoder;
 
-    public function __construct(ApiTokenRepository $apiTokenRepo)
+
+    public function __construct(ApiTokenRepository $apiTokenRepo, UserPasswordEncoderInterface $passwordEncoder)
     {
         $this->apiTokenRepo = $apiTokenRepo;
+        $this->passwordEncoder = $passwordEncoder;
     }
 
     public function supports(Request $request)
